Restructuring fourth pass
